Learning evidence page now shows assessments
This is still a prototype, in that matching the descriptions of assessments to how they are presented in a university's handbook is somewhat unclear.

e.g. see the entries for COSC220 and COSC320; in COSC320 the entire unit is shown as one assessment (even though in handbook it is separated) while in COSC220 the name of the "individual contribution" element doesn't match its handbook name